+++ airline-project/src/main/java/app/services/interfaces/FlightService.java
@@ -40,4 +40,11 @@ public interface FlightService {
                                                                     Date departureDate);
 
     void deleteFlightById(Long id);
+
+    Long getDistance(Flight flight);
+
+    public double parseLatitude(Airport airport);
+
+    public double parseLongitude(Airport airport);
+
 }